explorer/requirements.txt
streamlit>=1.30.0
streamlit-mermaid>=0.2.0
streamlit-agraph>=0.0.45
pyyaml>=6.0
python-frontmatter>=1.0.0
plotly>=5.18.0
pandas>=2.0.0
Pillow>=10.0.0
pygments>=2.17.0
whoosh>=2.7.4
